pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cloudinary>=1.44.1",
    "fastapi>=0.116.1",
    "fastapi-jwt[authlib]>=0.3.0",
    "fastapi-mail>=1.5.0",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pydantic[email]>=2.11.7",
    "pymongo[srv]>=4.13.2",
    "python-dotenv>=1.1.1",
    "python-slugify>=8.0.4",
    "uvicorn>=0.35.0",
]
